Add functionality to create new cards with answer options
In this commit, logic is added to create new cards with associated answer options in the application. The card creation process is now handled via a new method in the TestService, which takes in user information and a AddCardRequest. New classes to represent answer options and adding a card request are also introduced. Cases for transactions failing are handled with the addition of BeginTransactionAsync in the base repository. The TestController now includes a method to handle the card creation request.