Refactor Response DTOs to separate sensitive data
The DTO classes `CardResponse` and `AnswerOptionResponse` were removed and replaced by the more general `BaseCardResponse` and `BaseAnswerOptionResponse`, which exclude the `IsCorrect` field. This field is sensitive data that should not be made accessible to all routes. For routes where this data is necessary, we have created the `CardResponseWitCorrectAnswer` and the `AnswerOptionResponseWitCorrectAnswer` classes which extend the base classes and include the `IsCorrect` field. Adjustments have been made to the relevant services and controllers to use the new DTOs.